+++ src/collective/documentgenerator/tests/test_utils.py
@@ -181,41 +181,29 @@ class TestUtils(PODTemplateIntegrationTest):
 
         for renderer in (False, True):
             # convert to pdf
-            output_name, content = convert_file(odt_blob_file, filename.replace(".odt", ".pdf"), fmt="pdf",
-                                                renderer=renderer)
-            self.assertEqual(output_name, u"test_file.pdf")
+            content = convert_file(odt_blob_file, fmt="pdf", renderer=renderer)
             self.assertTrue(content.startswith("%PDF-"))
 
             # convert to odt
-            output_name, content = convert_file(odt_blob_file, filename.replace(".odt", ".odt"), fmt="odt",
-                                                renderer=renderer)
-            self.assertEqual(output_name, u"test_file.odt")
+            content = convert_file(odt_blob_file, fmt="odt", renderer=renderer)
             self.assertTrue(content.startswith("PK"))
             self.assertIn("mimetype", content)  # ODT files contain 'mimetype' in the zip
 
             # convert to docx
-            output_name, content = convert_file(odt_blob_file, filename.replace(".odt", ".docx"), fmt="docx",
-                                                renderer=renderer)
-            self.assertEqual(output_name, u"test_file.docx")
+            content = convert_file(odt_blob_file, fmt="docx", renderer=renderer)
             self.assertTrue(content.startswith("PK"))
             self.assertIn("[Content_Types].xml", content)  # DOCX files contain this file
 
             # convert to rtf
-            output_name, content = convert_file(odt_blob_file, filename.replace(".odt", ".rtf"), fmt="rtf",
-                                                renderer=renderer)
-            self.assertEqual(output_name, u"test_file.rtf")
+            content = convert_file(odt_blob_file, fmt="rtf", renderer=renderer)
             self.assertTrue(content.startswith("{\\rtf1"))
 
             # convert to txt
-            output_name, content = convert_file(odt_blob_file, filename.replace(".odt", ".txt"), fmt="txt",
-                                                renderer=renderer)
-            self.assertEqual(output_name, u"test_file.txt")
+            content = convert_file(odt_blob_file, fmt="txt", renderer=renderer)
             self.assertIn("Page 1\nPage 2\n", content)
 
             # convert to html
-            output_name, content = convert_file(odt_blob_file, filename.replace(".odt", ".html"), fmt="html",
-                                                renderer=renderer)
-            self.assertEqual(output_name, u"test_file.html")
+            content = convert_file(odt_blob_file, fmt="html", renderer=renderer)
             self.assertTrue(content.startswith("<!DOCTYPE html>"))
             self.assertIn("<html>", content)
             self.assertIn("<body", content)